# openapi.yaml
openapi: 3.0.3
info:
  title: Karton API
  description: >
    API za aplikaciju "Karton" — vođenje auto servisa. Backend je jedini autoritet
    za poslovna pravila (specifikacija-finalna.md §7). Nazivi na engleskom (odluka 27),
    poruke grešaka na srpskom. Sva vremena u UTC; poslovna logika Europe/Belgrade.

    CSRF ugovor: svaki odgovor postavlja/osvežava XSRF-TOKEN cookie (čitljiv iz JS);
    svaki non-GET zahtev MORA poslati header X-CSRF-Token sa tom vrednošću —
    neslaganje ili izostanak → 403 CSRF_FAILED. Važi i za /auth/login (token se
    dobija prethodnim GET pozivom, npr. /auth/me).

    Optimističko zaključavanje: verzionisani agregati su work_order, document,
    appointment i settings. Sve mutacione akcije nad njima nose "version" u telu;
    zastarela verzija → 409 VERSION_CONFLICT. Akcije bez version polja (copy,
    DELETE termina, izdavanje predračuna) backend izvršava u transakciji uz
    row-lock i atomsku proveru aktuelnog stanja.
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - cookieAuth: []
...
paths:
...
  /appointments:
    get:
      tags: [appointments]
      parameters:
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
        - { name: mechanicId, in: query, schema: { type: integer } }
        - $ref: '#/components/parameters/Query'
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Appointment' } } } } }
    post:
      tags: [appointments]
      summary: Novi termin — meka upozorenja 409 CONFIRMATION_REQUIRED (MECHANIC_BUSY, OUTSIDE_WORK_HOURS), blokiran dan 422 CALENDAR_BLOCKED
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AppointmentInput' } } } }
      responses:
        '201': { description: Zakazan, content: { application/json: { schema: { $ref: '#/components/schemas/Appointment' } } } }
        '409': { $ref: '#/components/responses/ConfirmationRequired' }
        '422': { $ref: '#/components/responses/BusinessRule' }
  /appointments/{id}:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    patch:
      tags: [appointments]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AppointmentInput' } } } }
      responses: { '200': { description: OK }, '409': { $ref: '#/components/responses/ConfirmationRequired' } }
    delete:
      tags: [appointments]
      summary: Fizičko brisanje — samo termin koji nije počeo (BR-29)
      responses: { '204': { description: Obrisan }, '422': { $ref: '#/components/responses/BusinessRule' } }
  /appointments/{id}/status:
    post:
      tags: [appointments]
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      summary: Statusna tranzicija (§6); completed→scheduled samo bez naloga (BR-28); uz completed opciono workOrderId
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status, version]
              properties:
                status: { $ref: '#/components/schemas/AppointmentStatus' }
                workOrderId: { type: integer, description: Vezivanje realizacije za nalog }
                reason: { type: string, description: Obavezan za admin korekcije }
                version: { type: integer }
      responses: { '200': { description: OK }, '409': { $ref: '#/components/responses/Conflict' }, '422': { $ref: '#/components/responses/BusinessRule' } }
  /calendar-blocks:
    get:
      tags: [appointments]
      responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/CalendarBlock' } } } } } }
    post:
      tags: [appointments]
      summary: Blokiraj dane (tvrda blokada za termine)
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CalendarBlockInput' } } } }
      responses: { '201': { description: Blokirano } }

  # ---------- Reports ----------
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: karton_session
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: integer } }
...
    Query: { name: q, in: query, schema: { type: string }, description: Case-insensitive, ignoriše razmake }
  responses:
...
    Conflict:
      description: VERSION_CONFLICT — optimističko zaključavanje
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    ConfirmationRequired:
      description: CONFIRMATION_REQUIRED — warnings lista; ponoviti sa confirmed=true
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    BusinessRule:
      description: 'Poslovno pravilo (422) — kod iz kataloga grešaka (spec §8)'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: PROFORMA_NOT_VALID }
        message: { type: string, example: 'Predračun je istekao i ne može se pretvoriti u račun.' }
        fields: { type: object, additionalProperties: { type: string } }
        warnings: { type: array, items: { type: string }, example: [MECHANIC_BUSY] }
        existingId: { type: integer, description: Kod duplikata — ID postojećeg entiteta }
...
    AppointmentStatus: { type: string, enum: [scheduled, completed, cancelled, no_show] }
...
    Customer:
      type: object
      properties:
        id: { type: integer }
        type: { type: string, enum: [individual, company] }
        name: { type: string }
        taxId: { type: string, nullable: true }
        address: { type: string }
        status: { type: string, enum: [active, archived] }
        contacts: { type: array, items: { $ref: '#/components/schemas/Contact' } }
...
    Contact:
      type: object
      properties:
        id: { type: integer }
        kind: { type: string, enum: [phone, email] }
        value: { type: string }
        isPrimary: { type: boolean }
...
    Vehicle:
      type: object
      properties:
        id: { type: integer }
        vin: { type: string }
        make: { type: string }
        model: { type: string }
        year: { type: integer }
        fuel: { type: string }
        note: { type: string }
        status: { type: string, enum: [active, archived] }
        currentPlate: { type: string, description: Izvedeno iz aktivnog zapisa }
        currentOwner: { $ref: '#/components/schemas/Customer' }
...
    Mechanic:
      type: object
      properties:
        id: { type: integer }
        fullName: { type: string }
        hiredOn: { type: string, format: date }
        hourlyRate: { type: number }
        specialty: { type: string, enum: [mechanical, electrical, other] }
        status: { type: string, enum: [active, inactive] }
...
    Appointment:
      type: object
      properties:
        id: { type: integer }
        date: { type: string, format: date }
        time: { type: string, example: '10:30' }
        durationMin: { type: integer }
        customer: { $ref: '#/components/schemas/Customer' }
        vehicle: { $ref: '#/components/schemas/Vehicle' }
        mechanic: { $ref: '#/components/schemas/Mechanic' }
        note: { type: string }
        status: { $ref: '#/components/schemas/AppointmentStatus' }
        workOrderId: { type: integer, nullable: true }
        remindersEnabled: { type: boolean }
        reminderStatus: { type: string, enum: [scheduled, processing, sent, failed], nullable: true, description: Status slanja — vidljiv na terminu }
        version: { type: integer }
    AppointmentInput:
      type: object
      properties:
        date: { type: string, format: date }
        time: { type: string }
        durationMin: { type: integer, default: 60 }
        customerId: { type: integer }
        vehicleId: { type: integer }
        mechanicId: { type: integer, nullable: true }
        note: { type: string }
        remindersEnabled: { type: boolean, description: Klijent bez emaila → 422 NO_CUSTOMER_EMAIL }
        confirmed: { type: boolean, default: false, description: Potvrda mekih upozorenja (BR-27) }
        version: { type: integer }
    CalendarBlock:
      type: object
      properties:
        id: { type: integer }
        fromDate: { type: string, format: date }
        toDate: { type: string, format: date }
        reason: { type: string }
    CalendarBlockInput:
      type: object
      required: [fromDate, toDate]
      properties:
        fromDate: { type: string, format: date }
        toDate: { type: string, format: date }
        reason: { type: string }
...